* Update roles and users api controllers

+++ Buzakov.App.FiveFilms/Buzakov.App.Services/RoleManagementService.cs
@@ -1,7 +1,9 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Data.Entity.Core;
 using System.Linq;
-using System.Runtime.InteropServices;
 using System.Web.Mvc;
+
 using Microsoft.AspNet.Identity.EntityFramework;
 
 using Buzakov.App.DataContext.EntityManager;
@@ -15,7 +17,6 @@ namespace Buzakov.App.Services
 
         private readonly IEntityManager _entityManager;
         private readonly RoleRepository _roleRepository;
-        
 
         public RoleManagementService( )
         {
@@ -28,14 +29,43 @@ namespace Buzakov.App.Services
             return _roleRepository.AsQueryable( ).ToList( );
         }
 
-        public IdentityRole Get(string id)
+        public IdentityRole Details( string id )
         {
-            return _roleRepository.Find(id);
+            if( id == null ) {
+                throw new ArgumentNullException("id");
+            }
+
+            var entity = _roleRepository.Find(id);
+            if( entity == null ) {
+                throw new ObjectNotFoundException( );
+            }
+
+            return entity;
         }
 
-        public void Delete(string id)
+        public IdentityRole Create( string name )
         {
+            if( name == null ) {
+                throw new ArgumentNullException("name");
+            }
+            var entity = _roleRepository.Insert(new IdentityRole { Name = name });
+
+            _entityManager.Commit( );
+
+            return entity;
+        }
+
+        public void Delete( string id )
+        {
+            if( id == null ) {
+                throw new ArgumentNullException("id");
+            }
+
             var entity = _roleRepository.Find(id);
+            if( entity == null ) {
+                throw new ObjectNotFoundException( );
+            }
+
             _roleRepository.Delete(entity);
 
             _entityManager.Commit( );
